// src/app/accounts/models/braintreeCard.ts
export class BraintreeCard {
    constructor() {

    }
     Address: string;
     Amount: number;
     CardNumber: string;
     City: string
     Comment1: string
     Comment2: string
     Country: number
     CurrencyCode: string
     CvvValue: string
     DoAuthorize: boolean
     EmailAddress: string
     ExpiryDate: string
     FirstName: string
     HomePhone: string
     IpAddress: string
     LastName: string
     OrderId: string
     State: string
     ZipCode: string
	 
}